Validate object ids

+++ src/controllers/user.js
@@ -26,6 +26,8 @@ export const authenticateUser = asyncHandler(async (req, res, next) => {
 });
 
 export const deleteUser = asyncHandler(async (req, res, next) => {
+  const user = await User.findById(req.params.userId);
+  if (!user) return res.status(404).json({ message: 'User not found' });
   await User.deleteOne({ _id: req.params.userId });
 
   res.status(200).json({ message: 'User deleted' });
